move graphQL code to 	crates/data_components/src/graphql/
