+++ crates/data_components/src/graphql/mod.rs
@@ -1,0 +1,59 @@
+/*
+Copyright 2024 The Spice.ai OSS Authors
+
+Licensed under the Apache License, Version 2.0 (the "License");
+you may not use this file except in compliance with the License.
+You may obtain a copy of the License at
+
+     https://www.apache.org/licenses/LICENSE-2.0
+
+Unless required by applicable law or agreed to in writing, software
+distributed under the License is distributed on an "AS IS" BASIS,
+WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+See the License for the specific language governing permissions and
+limitations under the License.
+*/
+
+use arrow::error::ArrowError;
+use snafu::Snafu;
+
+pub mod client;
+pub mod provider;
+
+#[derive(Debug, Snafu)]
+pub enum Error {
+    #[snafu(display("{source}"))]
+    ReqwestInternal { source: reqwest::Error },
+
+    #[snafu(display("HTTP {status}: {message}"))]
+    InvalidReqwestStatus {
+        status: reqwest::StatusCode,
+        message: String,
+    },
+
+    #[snafu(display("{source}"))]
+    ArrowInternal { source: ArrowError },
+
+    #[snafu(display("Invalid object access. {message}"))]
+    InvalidObjectAccess { message: String },
+
+    #[snafu(display(
+        r#"GraphQL Query Error:
+Details:
+- Error: {message}
+- Location: Line {line}, Column {column}
+- Query:
+
+{query}
+
+Please verify the syntax of your GraphQL query."#
+    ))]
+    InvalidGraphQLQuery {
+        message: String,
+        line: usize,
+        column: usize,
+        query: String,
+    },
+}
+
+pub type Result<T, E = Error> = std::result::Result<T, E>;
